# [MRG] Add clone_kernel option to make gaussian process models faster