--- a/doc/whats_new/v0.22.rst
+++ b/doc/whats_new/v0.22.rst
@@ -141,9 +141,6 @@ Changelog
   and :func:`gaussian_process.GaussianProcessRegressor.log_marginal_likelihood` now
   accept a ``clone_kernel=True`` keyword argument. When set to ``False``,
   the kernel attribute is modified, but may result in a performance improvement.
-  :func:`gaussian_process.GaussianProcessClassifier.fit` and
-  :func:`gaussian_process.GaussianProcessRegressor.fit` call
-  ``log_marginal_likelihood`` method with ``clone_kernel=False`` now.
   :pr:`14378` by :user:`Masashi Shibata <c-bata>`.
 
 - |API| From version 0.24 :meth:`Kernel.get_params` will raise an