--- a/sklearn/gaussian_process/gpc.py
+++ b/sklearn/gaussian_process/gpc.py
@@ -321,8 +321,8 @@ class _BinaryGaussianProcessClassifierLaplace(BaseEstimator):
             additionally. If True, theta must not be None.
 
         clone_kernel : bool, default=True
-            If True, the kernel attribute is copied. Please specify False if
-            you want to make this method call faster.
+            If True, the kernel attribute is copied. If False, the kernel
+            attribute is modified, but may result in a performance improvement.
 
         Returns
         -------
@@ -722,8 +722,8 @@ class GaussianProcessClassifier(BaseEstimator, ClassifierMixin):
             for non-binary classification. If True, theta must not be None.
 
         clone_kernel : bool, default=True
-            If True, the kernel attribute is copied. Please specify False if
-            you want to make this method call faster.
+            If True, the kernel attribute is copied. If False, the kernel
+            attribute is modified, but may result in a performance improvement.
 
         Returns
         -------