--- a/sklearn/gaussian_process/gpr.py
+++ b/sklearn/gaussian_process/gpr.py
@@ -404,7 +404,7 @@ class GaussianProcessRegressor(BaseEstimator, RegressorMixin,
             to the kernel hyperparameters at position theta is returned
             additionally. If True, theta must not be None.
 
-        clone_kernel : bool, default: True
+        clone_kernel : bool, default=True
             If True, the kernel attribute is copied. Please specify False if
             you want to make this method call faster.