#### Reference Issues/PRs
None

#### What does this implement/fix? Explain your changes.

**Problems**

I want to make GaussianProcessRegressor faster. I profiled `GaussianProcessRegressor#fit` using cProfile and generated the graph to find bottlenecks. The result is below:

![output](https://user-images.githubusercontent.com/5564044/61235963-c6cd5800-a771-11e9-8ee3-daf4600ba6ee.png)

This result shows `Kernel#clone_with_theta` accounts for 55.18% of the whole.

**Solutions**

Basically, I think we don't need to clone kernels. It seems to be enough that we just replace `theta` attribute. From my benchmark of [this gist](https://gist.github.com/c-bata/7eb7738c829c9858f4955dab5ac851e3), `GaussianProcessRegressor#fit` will be 26.4% faster by applying this patch.

*Before:*

```
$ python examples/gpr_bench.py 
elapsed: 3.620s
elapsed: 4.616s
elapsed: 3.892s
elapsed: 3.784s
elapsed: 3.908s
elapsed: 4.280s
elapsed: 2.884s
elapsed: 2.543s
elapsed: 2.502s
elapsed: 4.946s
score: 3.6975975036621094 0.7911395580877181
```

*After:*

```
$ python examples/gpr_bench.py 
elapsed: 2.642s
elapsed: 2.423s
elapsed: 2.921s
elapsed: 2.704s
elapsed: 2.856s
elapsed: 3.667s
elapsed: 2.737s
elapsed: 2.397s
elapsed: 2.292s
elapsed: 2.590s
score: 2.722917938232422 0.36802930790866245
```

#### Any other comments?

To confirm this PR doesn't break the logic, I generate the graph of model prediction results (I might need to design more suitable kernel functions).

| before | after (exactly the same with before's one) |
| ---  | --- |
| ![gpr_predict](https://user-images.githubusercontent.com/5564044/61236018-ef555200-a771-11e9-841e-aa6602ddaa7e.png) | ![gpr_predict2](https://user-images.githubusercontent.com/5564044/61236019-efede880-a771-11e9-8957-93a1b5d81ee7.png) |