@c-bata 2019-07-16
Thanks for reviewing! I'll apply this change to `gpc.py` 👍 

> It could help if you could also check that results before and after this change are identical not visually but with a numpy.testing.assert_allclose (even if we don't add it to tests).

I confirmed `numpy.testing.assert_allclose` doesn't raise AssertionError in following code:

```python
import numpy as np
import matplotlib.pyplot as plt
from sklearn.gaussian_process import kernels as sk_kern
from sklearn.gaussian_process import GaussianProcessRegressor


def objective(x):
    return x + 20 * np.sin(x)


def main():
    kernel = sk_kern.RBF(1.0, (1e-3, 1e3)) + sk_kern.ConstantKernel(1.0, (1e-3, 1e3))
    clf = GaussianProcessRegressor(
        kernel=kernel,
        alpha=1e-10,
        optimizer="fmin_l_bfgs_b",
        n_restarts_optimizer=20,
        normalize_y=True)

    np.random.seed(0)

    x_train = np.random.uniform(-20, 20, 200)
    y_train = objective(x_train) + np.random.normal(loc=0, scale=.1, size=x_train.shape)

    clf.fit(x_train.reshape(-1, 1), y_train)

    x_test = np.linspace(-20., 20., 200).reshape(-1, 1)
    pred_mean, pred_std = clf.predict(x_test, return_std=True)

    # save pred_mean and pred_std
    #np.save("pred_mean", pred_mean)
    #np.save("pred_std", pred_std)

    # assertion
    orig_mean = np.load("pred_mean.npy")
    orig_std = np.load("pred_std.npy")
    np.testing.assert_allclose(orig_mean, pred_mean)
    np.testing.assert_allclose(orig_std, pred_std)


if __name__ == '__main__':
    main()
```