@c-bata 2019-07-16
I added benchmark for GaussianProcessClassifier. The code and the results are below
https://gist.github.com/c-bata/2c7c8aa08852b5bc05bea37f33cd1962

And I also confirmed `np.testing.assert_allclose` doesn't raise AssertionError when asserting the results of prediction probabilities.
https://gist.github.com/c-bata/2c7c8aa08852b5bc05bea37f33cd1962#file-gpc_bench-py-L28-L29