@adrinjalali 2019-07-18
`log_marginal_likelihood()` is a public function, and you're changing the behavior of the function here.

This PR changes the output of the third line here:

```python
gpr.fit(X, y)
gpr.log_marginal_likelihood(theta=another_value)
gpr.log_marginal_likelihood()
```

You could maybe add a parameter like `clone_kernel=True` to `log_marginal_likelihood`, and pass `False` during `fit`, I guess.