@c-bata 2019-07-19
Thanks for reviewing, @adrinjalali !
To confirm whether the output of the third line is changed, I executed following script. But the output of the third line didn't be changed.
https://gist.github.com/c-bata/e6210a345a8431fbe293a725d1f4cc16

Could you tell me the case which the output of the third line will be changed?