@adrinjalali 2019-07-19
I haven't checked why the output doesn't change, but if it doesn't that's another issue with the code which we need to investigate. Your proposed patch looks good, please commit them here and we'll continue from there.