@c-bata 2019-07-30
Hi, @adrinjalali! Thanks for your review.
I pushed the commit to apply your comments :)